Reuse prepared statements
I made this change mostly because I like the organization better; I'd
rather carry around two Queries than the table name and the DB session.

I also added a benchmark test which I ran before and after the change to
be satisfied that there was no significant difference in performance.